+++ src/distributions.cpp
@@ -9,10 +9,10 @@
 
 #include "distributions.h"
 
-unsigned *rDirichletMultinomial(const unsigned count, const float *a,
+unsigned *rDirichletMultinomial(const unsigned count, const double *a,
                                 const unsigned d, std::mt19937 *engine) {
   // Draw p ~ Dirichlet(a)
-  float *p = rDirichlet(a, d, engine);
+  double *p = rDirichlet(a, d, engine);
   // Draw out ~ Multinomial(p)
   unsigned *out = rMultinomial(count, p, d, engine);
   // Cleanup and return
@@ -20,10 +20,10 @@ unsigned *rDirichletMultinomial(const unsigned count, const float *a,
   return out;
 }
 
-unsigned *rMultinomial(unsigned count, const float *p, const unsigned d,
+unsigned *rMultinomial(unsigned count, const double *p, const unsigned d,
                        std::mt19937 *engine) {
   unsigned *out = new unsigned[d];
-  float sum_ps = 1.0;
+  double sum_ps = 1.0;
   for (unsigned i = 0; i < d - 1; ++i) {
     // Draw from marginal binomial distribution.
     std::binomial_distribution<unsigned> b(count, p[i] / sum_ps);
@@ -36,13 +36,13 @@ unsigned *rMultinomial(unsigned count, const float *p, const unsigned d,
   return out;
 }
 
-float *rDirichlet(const float *a, const unsigned d, std::mt19937 *engine) {
-  float *gamma = new float[d];
-  float gamma_sum = 0.;
+double *rDirichlet(const double *a, const unsigned d, std::mt19937 *engine) {
+  double *gamma = new double[d];
+  double gamma_sum = 0.;
 
   // Sample the gamma variates for category i.
   for (unsigned i = 0; i < d; ++i) {
-    std::gamma_distribution<float> g(a[i]);
+    std::gamma_distribution<double> g(a[i]);
     gamma[i] = g(*engine);
     gamma_sum += gamma[i];
   }
